Integrate ElasticSearch in a Java EE6 stack
This is the first implementation of integrating ElasticSearch in a Java
EE 6 stack.
Injection into Entity Listeners is not yet supported [1] as the result
we need to use a GenericDao to notify ElasticSearch. Another solution is
to create/use a River [2] to pull data from the database.

The notification process use CDI Event and the listeners are responsible
for creating/updating/removing index.

The event type is not yet implemented (remove index).

[1] https://issues.jboss.org/browse/CDI-144
[2] http://www.elasticsearch.org/guide/reference/river/